new demos page; refactor catalog filters to make them reusabile
