+++ qdrant-landing/content/demo/_index.md
@@ -1,41 +1,11 @@
 ---
-title: Qdrant Demos and Tutorials
-description: Experience firsthand how Qdrant powers intelligent search, anomaly detection, and personalized recommendations, showcasing the full capabilities of vector search to revolutionize data exploration and insights.
-cards:
-  - id: 0
-    title: Semantic Search Demo - Startup Search
-    paragraphs:
-      - id: 0
-        content: This demo leverages a pre-trained SentenceTransformer model to perform semantic searches on startup descriptions, transforming them into vectors for the Qdrant engine.
-      - id: 1
-        content: Enter a query to see how neural search compares to traditional full-text search, with the option to toggle neural search on and off for direct comparison.
-    link:
-      text: View Demo
-      url: https://qdrant.to/semantic-search-demo
-  - id: 1
-    title: Semantic Search and Recommendations Demo - Food Discovery
-    paragraphs:
-      - id: 0
-        content: Explore personalized meal recommendations with our demo, using Delivery Service data. Like or dislike dish photos to refine suggestions based on visual appeal.
-      - id: 1
-        content: Filter options allow for restaurant selections within your delivery area, tailoring your dining experience to your preferences.
-    link:
-      text: View Demo
-      url: https://food-discovery.qdrant.tech/
-  - id: 2
-    title: Categorization Demo -<br> E-Commerce Products
-    paragraphs:
-      - id: 0
-        content: Discover the power of vector search in e-commerce through our demo. Simply input a product name and watch as our multi-language model intelligently categorizes it. The dots you see represent product clusters, highlighting our system's efficient categorization.
-    link:
-      text: View Demo
-      url: https://qdrant.to/extreme-classification-demo
-  - id: 3
-    title: Code Search Demo -<br> Explore Qdrant's Codebase
-    paragraphs:
-      - id: 0
-        content: Semantic search isn't just for natural language. By combining results from two models, qdrant is able to locate relevant code snippets down to the exact line.
-    link:
-      text: View Demo
-      url: https://code-search.qdrant.tech/
+title: Qdrant Demos
+description: Experience how Qdrant powers applications such as RAG, agentic AI, semantic/multimodal search, anomaly detection, and recommendations.
+build:
+  render: always
+cascade:
+  - build:
+      list: local
+      publishResources: false
+      render: never
 ---
