+++ qdrant-landing/content/demo/items/_index.md
@@ -1,0 +1,41 @@
+---
+sitemapExclude: true
+categories:
+  - Basics
+  - Search Engineering
+  - Operations
+  - Develop
+demos:
+  - id: semantic-search
+    title: Semantic Search Demo - Startup Search
+    description: This demo leverages a pre-trained SentenceTransformer model to perform semantic searches on startup descriptions, transforming them into vectors for the Qdrant engine. Enter a query to see how neural search compares to traditional full-text search.
+    category: Search Engineering
+    image: /img/demos/demo-0.png
+    link:
+      text: View Demo
+      url: https://qdrant.to/semantic-search-demo
+  - id: food-discovery
+    title: Semantic Search and Recommendations Demo - Food Discovery
+    description: Explore personalized meal recommendations with our demo, using Delivery Service data. Like or dislike dish photos to refine suggestions based on visual appeal.
+    category: Basics
+    image: /img/demos/demo-1.png
+    link:
+      text: View Demo
+      url: https://food-discovery.qdrant.tech/
+  - id: ecommerce-categorization
+    title: Categorization Demo - E-Commerce Products
+    description: Discover the power of vector search in e-commerce. Input a product name and watch as our multi-language model intelligently categorizes it into product clusters.
+    category: Basics
+    image: /img/demos/demo-2.png
+    link:
+      text: View Demo
+      url: https://qdrant.to/extreme-classification-demo
+  - id: code-search
+    title: Code Search Demo - Explore Qdrant's Codebase
+    description: Semantic search isn't just for natural language. By combining results from two models, Qdrant is able to locate relevant code snippets down to the exact line.
+    category: Develop
+    image: /img/demos/demo-3.png
+    link:
+      text: View Demo
+      url: https://code-search.qdrant.tech/
+---
