+++ qdrant-landing/themes/qdrant-2024/layouts/partials/customers/customers-case-studies.html
@@ -34,6 +34,9 @@
 <section
   class="customers-case-studies"
   data-customers-case-study
+  data-catalog-filters
+  data-filter-keys="industry,product,company_size,location,use_cases"
+  data-multi-value-keys="use_cases"
   data-batch-size="12"
 >
   <div class="container customers-case-studies__container">
