--- qdrant-landing/themes/qdrant-2024/layouts/partials/demo.html
@@ -1,26 +1,0 @@
-<section class="demos">
-  <div class="container">
-    <div class="demos__header mb-5 mb-md-7">
-      <h2 class="demos__title mx-auto mx-lg-0">{{ .Params.title }}</h2>
-      <p class="demos__subtitle mx-auto mx-lg-0">{{ .Params.description }}</p>
-    </div>
-    <div class="demos__cards row">
-      {{ range $index, $element := .Params.cards }}
-        <div class="col-12">
-          <div class="card demos__card">
-            <div class="demos__card-content">
-              <h6 class="demos__card-title">{{ .title | safeHTML }}</h6>
-              <div class="demos__card-description">
-                {{ range .paragraphs }}
-                  <p>{{ .content }}</p>
-                {{ end }}
-              </div>
-              <a class="link link_dark link_md align-self-start" href="{{ .link.url }}"> {{ .link.text }} </a>
-            </div>
-            <div class="demos__card-image" style="background-image: url(/img/demos/demo-{{ $index }}.png)"></div>
-          </div>
-        </div>
-      {{ end }}
-    </div>
-  </div>
-</section>
