+++ qdrant-landing/themes/qdrant-2024/layouts/partials/demo/demos-hero.html
@@ -1,0 +1,8 @@
+<section class="demos-hero">
+  <div class="container">
+    <div class="demos-hero__content">
+      <h1 class="demos-hero__title">{{ .Params.title }}</h1>
+      <p class="demos-hero__description">{{ .Params.description }}</p>
+    </div>
+  </div>
+</section>
